+++ mailhub-potluck/CHANGELOG.md
@@ -16,6 +16,7 @@
 * Fix postfix ldap settings again
 * Add obscure setting 'receive_override_options = no_address_mappings' in main.cf to allow delivery to 'username' (not 'username@domain') accounts which aren't local but are in virtual aliases
 * Permit authenticated and my networks instead of blocking at helo time
+* Add custom python-policyd-spf.conf
 
 ---